add args.json; save konect metadata table to csv; remove cache size filtering for graphs; change atoi to atol(l); modify igraph vector structs; add python-igraph connected components computation to use as verification of correctness for parallel slashburn;

+++ graph_preprocess/cuthill_mckee.cpp
@@ -37,10 +37,10 @@ int main(int argc, char *argv[]) {
 	while ((opt = getopt(argc, argv, "n:m:p:g:b:o:")) != -1) {
 		switch (opt) {
 			case 'n':
-				num_nodes = atoi(optarg);
+				num_nodes = atol(optarg);
 				break;
 			case 'm':
-				num_edges = atoi(optarg);
+				num_edges = atoll(optarg);
 				break;
 			case 'g':
 				input_path = optarg;
